+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- svgdx-server: Added support for the command line transform options available
+  to the `svgdx` binary, e.g. to set theme or initial variables from the command
+  line.
+
 ## [0.31.0 - 2026-07-15]
 
 - Added: text-loc can now be given a 'Length' (ratio as 10% or 3/4, or